backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
python-multipart>=0.0.20
openai>=1.58.0
chromadb>=0.6.0
pypdf>=5.1.0
pydantic>=2.11.0
